Tasks 15-23

+++ bookstore/books/views.py
@@ -1,8 +1,18 @@
 from django.shortcuts import render, redirect
 from .models import Book
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
+from django.contrib.auth import login 
+from .serializers import BookSerializer
+from rest_framework.decorators import api_view  
+from rest_framework.response import Response 
 
 def delete_book(request, id):
     book = Book.objects.get(id=id)
+
+    if book.posted_by != request.user:
+        from django.http import HttpResponseForbidden
+        return HttpResponseForbidden("You can only delete books you posted")
 
     if request.method == "POST":
         book.delete()
@@ -12,6 +22,10 @@ def delete_book(request, id):
 
 def edit_book(request, id):
     book = Book.objects.get(id=id)
+
+    if book.posted_by != request.user:
+        from django.http import HttpResponseForbidden
+        return HttpResponseForbidden("You can only edit books you posted")
 
     if request.method == 'POST':
         title = request.POST.get('title')
@@ -55,7 +69,7 @@ def edit_book(request, id):
 
     return render(request, 'books/edit.html', {'book': book})
 
-
+@login_required
 def add_book(request):
     if request.method == 'POST':
         title = request.POST.get('title')
@@ -109,3 +123,44 @@ def book_detail(request, id):
 def homepage(request):
     books = Book.objects.get_all_books()
     return render(request, 'books/home.html', {'books': books})
+
+
+def register(request):
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password1 = request.POST.get('password1')
+        password2 = request.POST.get('password2')
+
+        errors = []
+
+        if password1 != password2:
+            errors.append("Passwords don't match")
+        if User.objects.filter(username=username).exists():
+            errors.append("Username taken")
+        
+        if errors:
+            return render(request, 'books/register.html', {'errors': errors})
+
+        user = User.objects.create_user(username=username, password=password1)
+        login(request, user)  
+        return redirect('/')
+
+    return render(request, 'books/register.html')
+
+
+@api_view(['GET'])
+def api_get_books(request):
+    """Get all books as JSON"""
+    books = Book.objects.all()
+    serializer = BookSerializer(books, many=True)
+    return Response(serializer.data)
+
+@api_view(['GET'])
+def api_get_book(request, id):
+    """Get single book by ID as JSON"""
+    try:
+        book = Book.objects.get(id=id)
+        serializer = BookSerializer(book)
+        return Response(serializer.data)
+    except Book.DoesNotExist:
+        return Response({'error': 'Book not found'}, status=404)
